+++ NINA.Equipment/SDK/CameraSDKs/SVBonySDK/SVBonySDK.cs
@@ -483,7 +483,7 @@ namespace NINA.Equipment.SDK.CameraSDKs.SVBonySDK {
             return false;
         }
 
-        public bool SetDewHeater(bool onoff) {
+        public bool SetDewHeater(int power) {
             throw new NotImplementedException();
         }
 
@@ -507,6 +507,18 @@ namespace NINA.Equipment.SDK.CameraSDKs.SVBonySDK {
             return 0;
         }
         public void SetReadoutMode(int modeIndex) {
+        }        
+
+        public bool HasAdjustableFan() {
+            return false;
+        }
+
+        public bool SetFanPercentage(int fanPercentage) {
+            throw new NotImplementedException();
+        }
+
+        public int GetFanPercentage() {
+            throw new NotImplementedException();
         }
     }
 }