+++ NINA/View/Equipment/Camera/CameraTemplateSelector.cs
@@ -28,6 +28,7 @@ namespace NINA.View.Equipment {
         public DataTemplate Canon { get; set; }
         public DataTemplate Atik { get; set; }
         public DataTemplate Zwo { get; set; }
+        public DataTemplate Generic { get; set; }
         public DataTemplate FailedToLoadTemplate { get; set; }
 
         public string Postfix { get; set; }
@@ -45,6 +46,8 @@ namespace NINA.View.Equipment {
                 return Atik;
             } else if (item is ASICamera) {
                 return Zwo;
+            } else if (item is GenericCamera) {
+                return Generic;
             } else {
                 var templateKey = item?.GetType().FullName + Postfix;
                 if (item != null && Application.Current.Resources.Contains(templateKey)) {