+++ RELEASE_NOTES.md
@@ -52,6 +52,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 
 ### PlayerOne
 - Added temperature control for cooled camera series
+- Cooled cameras can now control dew heater and fan strength
 - Native driver for Player One Filter Wheels is now available
 
 # Version 2.1